+++ num_formatter.h
@@ -13,7 +13,7 @@ namespace tc
     class NumFormatter {
     public:
 
-        static std::string FormatStorageSize(int bytes);
+        static std::string FormatStorageSize(int64_t bytes);
         static std::string FormatTime(uint64_t timestamp);
 
     };
